Move Meta component into Meta directory
