+++ components/Page.js
@@ -1,31 +1,33 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { withRouter } from 'next/router';
+
 import Header from './Header/Header';
 import Footer from './Footer/Footer';
-import Meta from './Meta';
+import Meta from './Meta/Meta';
 import { capitalizeFirst } from '../lib/utils';
 
-const Page = props => {
-  const { router } = props;
+const Page = ( { children, router } ) => {
   const { pathname } = router;
 
   let path = pathname.substr( router.pathname.lastIndexOf( '/' ) + 1 );
+
   // if on a dynamic page, i.e. admin/package/[id] then get the page title
   // from one level up, 'package'
   if ( path.indexOf( '[' ) !== -1 ) {
     const parts = pathname.split( '/' );
+
     path = parts[parts.length - 2];
   }
   const title = path ? `${capitalizeFirst( path )} | Content Commons` : 'Content Commons';
-  const bodyCls = ( pathname === '/' ) ? '' : 'ui container inside';
+  const bodyCls = pathname === '/' ? '' : 'ui container inside';
 
   return (
     <div style={ { position: 'relative', minHeight: '100vh' } }>
       <Meta title={ title } />
       <Header />
       <main className={ bodyCls }>
-        { props.children }
+        { children }
       </main>
       <Footer />
     </div>
@@ -35,7 +37,7 @@ const Page = props => {
 
 Page.propTypes = {
   children: PropTypes.node,
-  router: PropTypes.object
+  router: PropTypes.object,
 };
